backend/pyproject.toml
[tool.poetry]
name = "bicycle-rent"
version = "0.1.0"
description = ""
authors = ["Smitona <jarofsmetana@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
argon2-cffi = "^23.1.0"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.2"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"

[tool.isort]
line_length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
